requirements.txt
sentinelsat==1.1.1
python-dotenv==0.20.0
numpy
pillow
pandas
matplotlib
filelock